Faster inference when using only boxes
