+++ model_helpers/prepare_positionnet.py
@@ -9,18 +9,18 @@ def get_positionnet_default_params():
         "mid_dim": 3072,
         "out_dim": 768,
         "test_drop_boxes": False,
-        "test_drop_masks": True,
+        "test_drop_masks": False,
         "test_drop_points": False,
-        "test_drop_scribbles": True,
+        "test_drop_scribbles": False,
         "train_add_boxes": True,
-        "train_add_masks": True,
-        "train_add_points": True,
-        "train_add_scribbles": True,
+        "train_add_masks": False,
+        "train_add_points": False,
+        "train_add_scribbles": False,
         "use_seperate_tokenizer": True,
     }
 
 
 def prepare_positionnet(checkpoint, params) -> torch.nn.Module:
     model = UniFusion(**params)
-    model.load_state_dict(checkpoint)
+    model.load_state_dict(checkpoint, strict=False)
     return model
