+++ modules/convnext.py
@@ -173,49 +173,49 @@ def convnext_tiny(pretrained=False, in_22k=False, **kwargs):
     return model
 
 
-@register_model
-def convnext_small(pretrained=False, in_22k=False, **kwargs):
-    model = ConvNeXt(depths=[3, 3, 27, 3], dims=[96, 192, 384, 768], **kwargs)
-    if pretrained:
-        url = model_urls['convnext_small_22k'] if in_22k else model_urls['convnext_small_1k']
-        checkpoint = torch.hub.load_state_dict_from_url(
-            url=url, map_location="cpu")
-        model.load_state_dict(checkpoint["model"])
-    return model
+# @register_model
+# def convnext_small(pretrained=False, in_22k=False, **kwargs):
+#     model = ConvNeXt(depths=[3, 3, 27, 3], dims=[96, 192, 384, 768], **kwargs)
+#     if pretrained:
+#         url = model_urls['convnext_small_22k'] if in_22k else model_urls['convnext_small_1k']
+#         checkpoint = torch.hub.load_state_dict_from_url(
+#             url=url, map_location="cpu")
+#         model.load_state_dict(checkpoint["model"])
+#     return model
 
 
-@register_model
-def convnext_base(pretrained=False, in_22k=False, **kwargs):
-    model = ConvNeXt(depths=[3, 3, 27, 3], dims=[
-                     128, 256, 512, 1024], **kwargs)
-    if pretrained:
-        url = model_urls['convnext_base_22k'] if in_22k else model_urls['convnext_base_1k']
-        checkpoint = torch.hub.load_state_dict_from_url(
-            url=url, map_location="cpu")
-        model.load_state_dict(checkpoint["model"])
-    return model
+# @register_model
+# def convnext_base(pretrained=False, in_22k=False, **kwargs):
+#     model = ConvNeXt(depths=[3, 3, 27, 3], dims=[
+#                      128, 256, 512, 1024], **kwargs)
+#     if pretrained:
+#         url = model_urls['convnext_base_22k'] if in_22k else model_urls['convnext_base_1k']
+#         checkpoint = torch.hub.load_state_dict_from_url(
+#             url=url, map_location="cpu")
+#         model.load_state_dict(checkpoint["model"])
+#     return model
 
 
-@register_model
-def convnext_large(pretrained=False, in_22k=False, **kwargs):
-    model = ConvNeXt(depths=[3, 3, 27, 3], dims=[
-                     192, 384, 768, 1536], **kwargs)
-    if pretrained:
-        url = model_urls['convnext_large_22k'] if in_22k else model_urls['convnext_large_1k']
-        checkpoint = torch.hub.load_state_dict_from_url(
-            url=url, map_location="cpu")
-        model.load_state_dict(checkpoint["model"])
-    return model
+# @register_model
+# def convnext_large(pretrained=False, in_22k=False, **kwargs):
+#     model = ConvNeXt(depths=[3, 3, 27, 3], dims=[
+#                      192, 384, 768, 1536], **kwargs)
+#     if pretrained:
+#         url = model_urls['convnext_large_22k'] if in_22k else model_urls['convnext_large_1k']
+#         checkpoint = torch.hub.load_state_dict_from_url(
+#             url=url, map_location="cpu")
+#         model.load_state_dict(checkpoint["model"])
+#     return model
 
 
-@register_model
-def convnext_xlarge(pretrained=False, in_22k=False, **kwargs):
-    model = ConvNeXt(depths=[3, 3, 27, 3], dims=[
-                     256, 512, 1024, 2048], **kwargs)
-    if pretrained:
-        assert in_22k, "only ImageNet-22K pre-trained ConvNeXt-XL is available; please set in_22k=True"
-        url = model_urls['convnext_xlarge_22k']
-        checkpoint = torch.hub.load_state_dict_from_url(
-            url=url, map_location="cpu")
-        model.load_state_dict(checkpoint["model"])
-    return model
+# @register_model
+# def convnext_xlarge(pretrained=False, in_22k=False, **kwargs):
+#     model = ConvNeXt(depths=[3, 3, 27, 3], dims=[
+#                      256, 512, 1024, 2048], **kwargs)
+#     if pretrained:
+#         assert in_22k, "only ImageNet-22K pre-trained ConvNeXt-XL is available; please set in_22k=True"
+#         url = model_urls['convnext_xlarge_22k']
+#         checkpoint = torch.hub.load_state_dict_from_url(
+#             url=url, map_location="cpu")
+#         model.load_state_dict(checkpoint["model"])
+#     return model
